Add Auth to Stable Diffusion
